refactor: 패키지 구조 일관성을 위한 정리
